+++ NEWS.md
@@ -23,7 +23,7 @@ __Support for timeline added__
 - `auritus` theme added.
 - `bar.width` argument renamed to `bar_width` in `e_histogram`.
 - `e_map_3d` and `e_map` correctly assign `name` argument.
-- Axis and tooltip formatters helpers added `e_tooltip_item_formatter` & `e_axis_formatter`, thanks to [Artem Klevtsov](https://github.com/artemklevtsov)
+- Axis and tooltip formatters helpers added `e_tooltip_item_formatter` & `e_axis_formatter`, thanks to [Artem Klevtsov](https://github.com/artemklevtsov).
 
 # echarts4r 0.2.0